Move learn components into app/ui/learn, extract Skill component, wire props
